app/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
torch
